Index type issue

+++ components/media/geometry/sources/binmesh/geometry_binmesh_loader.cpp
@@ -16,7 +16,7 @@ using namespace common;
  *
  * vertex_buffer_desc format: array<size_t> vertex_weigth_stream_id (0-1); array<size_t> vertex_stream_id;
  *
- * index_buffer format: array<size_t> indices
+ * index_buffer format: array<unsigned int> indices
  *
  * mesh format: array<char> name; array<size_t> index_buffer (0-1); array<size_t> vertex_buffers; array<primitive> primitives;
  *
@@ -217,7 +217,7 @@ class BinMeshLibraryLoader
 
         //чтение индексов
 
-      file_read (input_file, ib.Data (), indices_count * sizeof (size_t));
+      file_read (input_file, ib.Data (), indices_count * sizeof (unsigned int));
 
         //регистрация потока
